* commit 'd9f3aa58d7e651dfbe9b395086040887d32d5a98':
  Fix removeTarget to remove by ID instead of index.

+++ core/java/android/transition/Transition.java
@@ -1046,7 +1046,7 @@ public abstract class Transition implements Cloneable {
      */
     public Transition removeTarget(int targetId) {
         if (targetId > 0) {
-            mTargetIds.remove(targetId);
+            mTargetIds.remove((Integer)targetId);
         }
         return this;
     }
